requirements.txt
absl-py==2.2.2
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.52.2
anyio==4.8.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
bert-score==0.3.13
blinker==1.9.0
blis==1.3.0
build==1.2.2.post1
cachetools==5.5.1
catalogue==2.0.10
cbor==1.0.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.8
cloudpathlib==0.21.0
cohere==5.15.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.3.2
cryptography==44.0.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.1
dacite==1.9.2
dataclasses-json==0.6.7
datasets==3.5.0
de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.8.0/de_core_news_sm-3.8.0-py3-none-any.whl#sha256=fec69fec52b1780f2d269d5af7582a5e28028738bd3190532459aeb473bfa3e7
debugpy==1.8.12
decorator==5.1.1
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
durationpy==0.9
emoji==2.14.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.11.0
fastapi==0.115.8
fastavro==1.11.1
fastjsonschema==2.21.1
filelock==3.17.0
filetype==1.2.0
FlagEmbedding==1.3.5
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.58.2
frozenlist==1.5.0
fschat==0.2.36
fsspec==2024.12.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0
google-api-python-client==2.171.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.72.1
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.1.2
html2text==2025.4.15
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
ijson==3.4.0
import-ipynb==0.2
importlib_metadata==8.5.0
importlib_resources==6.5.2
importnb==2023.11.1
inscriptis==2.6.0
ipykernel==6.29.5
ipynb==0.5.1
ipython==8.32.0
ir_datasets==0.5.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
json_repair==0.39.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
kubernetes==32.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-experimental==0.3.4
langchain-openai==0.3.5
langchain-text-splitters==0.3.6
langcodes==3.5.0
langdetect==1.0.9
langsmith==0.1.147
language_data==1.3.0
latex2mathml==3.78.0
lxml==5.3.1
lz4==4.4.4
marisa-trie==1.2.1
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
narwhals==1.46.0
nbformat==5.10.4
nbimporter==0.3.4
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
nh3==0.2.21
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
openai==1.63.0
openpyxl==3.1.5
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
peft==0.15.2
pillow==11.2.1
platformdirs==4.3.6
posthog==3.13.0
preshed==3.0.9
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.3
psutil==7.0.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyjnius==1.6.1
pyparsing==3.2.3
pypdf==5.3.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyserini==0.44.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.8
python-magic==0.4.27
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
ragas==0.2.14
rank-bm25==0.2.2
rank-llm==0.21.0
RapidFuzz==3.12.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
sentence-transformers==4.1.0
sentencepiece==0.2.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
squarify==0.4.4
srsly==2.5.1
stack-data==0.6.3
starlette==0.45.3
streamlit==1.46.1
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
thinc==8.3.6
threadpoolctl==3.6.0
tiktoken==0.7.0
together==1.5.8
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.7.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trec-car-tools==2.6
typer==0.15.1
types-requests==2.32.0.20250515
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
unlzw3==0.2.3
unstructured==0.14.4
unstructured-client==0.30.0
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wordcloud==1.9.4
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zlib-state==0.1.9
